Backport some ImapStreamParser fixes from KIMAP, fix multi literal support
Backport a crashfix in readUntilCommandEnd() and support for client-side use of the
stream parser (i.e. for cases when literal parts are always part of the response, so
no need to request them).

Fix support for multiple literal parts in readUntilCommandEnd(), so that the method
really reads all literal parts, even when they are separated by CRLF.