Create a per-type ChangeNotification Protocol command
Different entity types have different data, so sharing a single
structure to describe changes in all entities is limiting. Having
a single structure made sense when we only had Items and Collections
which share most of the attributes and it was also easier to expose
this via DBus.

Now we are no longer limited by DBus so we can afford having a
different Protocol command for each notification type. This will be
most important when we introduce SubscriberChangeNotification, as
subscriber changes would otherwise be hard to map onto the legacy
ChangeNotification and is also the first step towards Notification
Payloads.